backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
python-multipart==0.0.9
google-generativeai==0.7.2
